refactor(webui): drop the voice surface the WAV re-encode made dead
Three simplifications, each a leftover from before the client-side WAV
conversion landed:

- `session.voice` no longer advertises a format list. The recorder picks its
  own container and the composer uploads exactly one format, so `accept` was a
  contract nothing read — and one that had already gone stale (it still listed
  the webm/mp4 the endpoint rejects). `VoiceCapabilities` collapses into
  `VoiceClipBudget`, serialized directly; `voice_capabilities()` and
  `audio_mime_types()` are gone. The server-side registry check in
  `DecodeVoiceClip` was always the real authority and is untouched.
- `RebornServices::transcription_available()` is removed: production reads the
  gate from composition through the serve config, so the accessor had exactly
  one caller, a test asserting it.
- `voice-encode.ts` folds into `voice-recorder.ts`. Both were lazy, in the same
  chunk, with the same lifetime, so the module boundary bought nothing.

Net -92 lines. Bundle is marginally smaller (222.6 -> 222.5 KB); one test
replaces the two that asserted the removed list, pinning that the single
upload format stays registry-supported.
