+++ crates/domains/ironclaw_attachments/src/lib.rs
@@ -29,8 +29,7 @@ pub const WORKSPACE_ALIAS: &str = "/workspace";
 
 pub use budgets::{
     AttachmentBudgets, AttachmentCapabilities, DEFAULT_ATTACHMENT_BUDGETS,
-    DEFAULT_VOICE_CLIP_BUDGET, VoiceCapabilities, VoiceClipBudget, attachment_capabilities,
-    audio_mime_types, voice_capabilities,
+    DEFAULT_VOICE_CLIP_BUDGET, VoiceClipBudget, attachment_capabilities,
 };
 pub use inbound::land_inbound_attachments;
 pub use landing::{
