+++ crates/product/ironclaw_webui/frontend/src/pages/chat/components/chat-input.test.ts
@@ -39,7 +39,7 @@ function renderChatInput({ sendDisabled = true, statusText = "" } = {}) {
     // tests opt in by overriding `useVoiceConfig`.
     useVoiceConfig: () => ({
       enabled: false,
-      limits: { accept: [], maxBytes: 1024, maxDurationSecs: 300 },
+      limits: { maxBytes: 1024, maxDurationSecs: 300 },
     }),
     useVoiceInput: () => ({
       status: "idle",
